chore: rename login and setting folder
